+++ src/components/ChannelList/__tests__/ChannelList.test.js
@@ -176,6 +176,37 @@ describe('ChannelList', () => {
     });
   });
 
+  it('should only show filtered channels when a filter function prop is provided', async () => {
+    const filteredChannel = generateChannel({ channel: { type: 'filtered' } });
+
+    const customFilterFunction = (channels) => {
+      return channels.filter((channel) => channel.type === 'filtered');
+    };
+
+    const props = {
+      filters: {},
+      Preview: ChannelPreviewComponent,
+      List: ChannelListComponent,
+      channelRenderFilterFn: customFilterFunction,
+    };
+
+    useMockedApis(chatClientUthred, [
+      queryChannelsApi([filteredChannel, testChannel1]),
+    ]);
+
+    const { getByRole, queryAllByRole } = render(
+      <Chat client={chatClientUthred}>
+        <ChannelList {...props} />
+      </Chat>,
+    );
+
+    // Wait for list of channels to load in DOM.
+    await waitFor(() => {
+      expect(getByRole('list')).toBeInTheDocument();
+      expect(queryAllByRole('listitem')).toHaveLength(1);
+    });
+  });
+
   it('should render `LoadingErrorIndicator` when queryChannels api throws error', async () => {
     useMockedApis(chatClientUthred, [erroredGetApi()]);
     jest.spyOn(console, 'warn').mockImplementationOnce(() => null);