+++ CHANGELOG.md
@@ -1,10 +1,23 @@
 # Changelog
 
+## [3.5.3](https://github.com/GetStream/stream-chat-react/releases/tag/v3.5.3) 2021-01-26
+
+### Feature
+
+- `ChannelList` accepts an optional prop to filter/sort channels prior to render [#678](https://github.com/GetStream/stream-chat-react/pull/678)
+
+## [3.5.2](https://github.com/GetStream/stream-chat-react/releases/tag/v3.5.2) 2021-01-21
+
+### Fix
+
+- Handle emoji picker event listener race condition [#675](https://github.com/GetStream/stream-chat-react/pull/675)
+- Adjust style for only emoji messages [#676](https://github.com/GetStream/stream-chat-react/pull/676)
+
 ## [3.5.1](https://github.com/GetStream/stream-chat-react/releases/tag/v3.5.1) 2021-01-19
 
 ### Fix
 
-- Upload PSD attachments as file [#67](https://github.com/GetStream/stream-chat-react/pull/673)
+- Upload PSD attachments as file [#673](https://github.com/GetStream/stream-chat-react/pull/673)
 
 ### Chore